refactor: 필요없는 파일, 코드 날림
