--- daeun.txt
@@ -1,39 +1,0 @@
-* use client 주의
-
-<InputField>
-import { useState } from "react";
-import InputField from "@/components/ui/InputField";
-
-export default function Home() {
-  const [name, setName] = useState("");
-  const [error, setError] = useState("");
-
-  const MAX_LENGTH = 30;
-
-  const handleChange = (e) => {
-    const value = e.target.value;
-    setName(value);
-
-    if (value.length > MAX_LENGTH) {
-      setError(`포토카드 이름은 최대 ${MAX_LENGTH}자까지 입력 가능합니다.`);
-    } else {
-      setError("");
-    }
-  };
-
-  return (
-    <div className="p-4 bg-black min-h-screen">
-      <InputField
-        label="포토카드 이름"
-        placeholder="포토카드 이름을 입력해 주세요"
-        value={name}
-        onChange={handleChange}
-        name="photocardName"
-        labelClassName="text-base font-bold"
-        error={error}
-      />
-    </div>
-  );
-}
-
-// InputField만 회원가입, 로그인 외에도 쓰일 곳이 있어서 이 컴포넌트만 txt로 드렸습니다
