+++ src/app/home/page.jsx
@@ -1,7 +1,7 @@
 import SellButton from "@/components/ui/SellButton";
 import BaseCardsSection from "@/components/HompageComponents/BaseCardsSection";
-import RandomPoint from "@/components/RandomPoint";
-import RandomPointHomeTrigger from "@/components/RandomPointHomeTrigger";
+import RandomPoint from "@/components/point/RandomPoint";
+import RandomPointHomeTrigger from "@/components/point/RandomPointHomeTrigger";
 import NotiModal from "@/components/modal/NotiModal";
 
 export const metadata = {
@@ -13,13 +13,13 @@ export const metadata = {
     type: "website",
     images: ["https://6-favorite-photo-4team-fe.vercel.app/contents.png"],
   },
-    twitter: {
+  twitter: {
     card: "summary_large_image",
     title: "최악의 포토",
     description: "가장 이상한 포토카드를 사고파는 곳, 지금 바로 시작해보세요!",
-    site: "@your_twitter", 
+    site: "@your_twitter",
     creator: "나야 나~",
-    images: ["https://6-favorite-photo-4team-fe.vercel.app/contents.png"], 
+    images: ["https://6-favorite-photo-4team-fe.vercel.app/contents.png"],
   },
 };
