+++ src/components/point/RandomPoint.jsx
