+++ src/components/point/RandomPointHomeTrigger.jsx
