--- src/delay.js
@@ -1,7 +1,0 @@
-export function delay(count) {
-    return new Promise((resolve) => {
-        setTimeout(() => {
-            resolve("Loading...")
-        },count)
-    })
-}
